Added tables connections and funcinality to show person and parking spaces together. SQL error 19 contraint foreign key still not resolved

+++ Parking_Lot_Reservation/Controllers/ParkingSpaceController.cs
@@ -28,11 +28,14 @@ namespace Parking_Lot_Reservation.Controllers
             var parkingSpaceAdd = new ParkingSpaceModel
             {
                 IsReserved = parkingSpaceDTO.IsReserved,
-                HasCharger = parkingSpaceDTO.HasCharger
+                HasCharger = parkingSpaceDTO.HasCharger,
             };
 
-            _ = await _dbContext.ParkingSpaces.AddAsync(parkingSpaceAdd);
-            _ = await _dbContext.SaveChangesAsync();
+            if (ModelState.IsValid)
+            {
+                _ = await _dbContext.ParkingSpaces.AddAsync(parkingSpaceAdd);
+                _ = await _dbContext.SaveChangesAsync();
+            }
 
             return parkingSpaceAdd;
         }
@@ -55,17 +58,22 @@ namespace Parking_Lot_Reservation.Controllers
         [HttpDelete("removeParkingSpace")]
         public async Task<ActionResult<ParkingSpaceDTO>> Remove(int id)
         {
-            var parkingSpace = await _dbContext.ParkingSpaces.FindAsync(id);
+            if (id < 0)
+            {
+                return new BadRequestObjectResult("Id is incorrect");
+            }
+
+            var parkingSpace = await _dbContext.ParkingSpaces.FirstOrDefaultAsync(parking => parking.ParkingId.Equals(id));
 
             if (parkingSpace is null)
             {
-                return new BadRequestResult();
+                return new BadRequestObjectResult("Parking space is null");
             }
 
-            _ = _dbContext.ParkingSpaces.Remove(parkingSpace);
+            _ =_dbContext.ParkingSpaces.Remove(parkingSpace);
             _ = await _dbContext.SaveChangesAsync();
 
-            return new ParkingSpaceDTO { HasCharger = parkingSpace.HasCharger, IsReserved = parkingSpace.IsReserved };
+            return new OkResult();
         }
     }
 }
